+++ storage/pebble/operations/codes.go
@@ -1,0 +1,5 @@
+package operations
+
+const (
+	codeChunkDataPack = 100
+)
