+++ storage/pebble/operations/common.go
@@ -1,0 +1,55 @@
+package operations
+
+import (
+	"errors"
+
+	"github.com/cockroachdb/pebble"
+	"github.com/vmihailenco/msgpack"
+
+	"github.com/onflow/flow-go/model/flow"
+	"github.com/onflow/flow-go/module/irrecoverable"
+	"github.com/onflow/flow-go/storage"
+)
+
+func insert(key []byte, val interface{}) func(pebble.Writer) error {
+	return func(w pebble.Writer) error {
+		value, err := msgpack.Marshal(val)
+		if err != nil {
+			return irrecoverable.NewExceptionf("failed to encode value: %w", err)
+		}
+
+		err = w.Set(key, value, nil)
+		if err != nil {
+			return irrecoverable.NewExceptionf("failed to store data: %w", err)
+		}
+
+		return nil
+	}
+}
+
+func retrieve(key []byte, sc interface{}) func(r pebble.Reader) error {
+	return func(r pebble.Reader) error {
+		val, closer, err := r.Get(key)
+		if err != nil {
+			return convertNotFoundError(err)
+		}
+		defer closer.Close()
+
+		err = msgpack.Unmarshal(val, &sc)
+		if err != nil {
+			return irrecoverable.NewExceptionf("failed to decode value: %w", err)
+		}
+		return nil
+	}
+}
+
+func makeKey(prefix byte, identifier flow.Identifier) []byte {
+	return append([]byte{prefix}, identifier[:]...)
+}
+
+func convertNotFoundError(err error) error {
+	if errors.Is(err, pebble.ErrNotFound) {
+		return storage.ErrNotFound
+	}
+	return err
+}
